SelectedNodeId теперь централизованно передается через NodeStore,в Dao перед вызовом Update и Delete вызывается проверка findById

+++ src/com/example/tree_rumyancev/server/dao/impl/TreeDaoMyBatisImpl.java
@@ -41,6 +41,9 @@ public class TreeDaoMyBatisImpl implements TreeDao {
 	public void update(Node node) {
 		try (SqlSession session = MyBatisUtil.getSqlSessionFactory().openSession()) {
 			TreeMapper mapper = session.getMapper(TreeMapper.class);
+			if(findById(node.getId())== null) {
+				return;
+			}
 			mapper.update(node);
 			session.commit();
 		}
@@ -50,6 +53,9 @@ public class TreeDaoMyBatisImpl implements TreeDao {
 	public void delete(Long id) {
 		try (SqlSession session = MyBatisUtil.getSqlSessionFactory().openSession()) {
 			TreeMapper mapper = session.getMapper(TreeMapper.class);
+			if(findById(id)== null) {
+				return;
+			}
 			mapper.delete(id);
 			session.commit();
 		}
